requirements.txt
tensorflow>=2.13.0
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
gradio==4.8.0
pandas==2.1.4
numpy==1.26.2
jupyter==1.0.0